Add support for Bukkit-based 1.8.8 servers
This patch does not modify any other platform implementations such as Fabric
or Forge, this is purely for servers running 1.8.8 which wish to support
clients using the mod on newer clients. The already-existing
ViaVersionCompatibility class handles mapping the channel name so 1.13+ clients
can understand the plugin channel name.

Usages of Bukkit's 'NamespacedKey' class were changed to Adventure's 'Key',
as this class was not present in 1.8.8.